move mine_vanity_key to pow in prep for refactor
