+++ test/test_key.py
@@ -1,5 +1,5 @@
 import pytest
-from nostr.key import PrivateKey, mine_vanity_key
+from nostr.key import PrivateKey
 
 
 def test_eq_true():
@@ -22,21 +22,3 @@ def test_from_nsec():
     pk1 = PrivateKey()
     pk2 = PrivateKey.from_nsec(pk1.bech32())
     assert pk1.raw_secret == pk2.raw_secret
-
-
-def test_mine_vanity_key():
-    """ test vanity key mining """
-    pattern = '23'
-
-    # mine a valid pattern as prefix
-    sk = mine_vanity_key(prefix=pattern)
-    sk.public_key.bech32()
-    assert sk.public_key.bech32().startswith(f'npub1{pattern}')
-
-    # mine a valid pattern as suffix
-    sk = mine_vanity_key(suffix=pattern)
-    assert sk.public_key.bech32().endswith(pattern)
-
-    # mine an invalid pattern
-    with pytest.raises(ValueError) as e:
-        mine_vanity_key(prefix='1')
